requirements.txt
Flask==2.1.3
Flask-SocketIO==5.1.0
requests==2.25.1
python-socketio==5.3.0 
Werkzeug==2.0.3